+++ src/listeners/ready.js
@@ -12,6 +12,7 @@ class ReadyListener extends Listener {
     run(client) {
         client.MutedMembers.sync();
         client.TimeZones.sync();
+        client.FagBucks.sync();
         const { username, id } = client.user;
         this.container.logger.info(`Successfully logged in as ${username} (${id})!`);
     }